edit expected prediction raster
